+++ test/test_helper.rb
@@ -16,13 +16,17 @@ def assert_string_array(strings, replace: nil, ignore_nbsp: true)
 	strings.each do |i|
 		f = Lightgraf.fix(
 				i[0],
-				html_encode: i[2]&.dig(:html_encode) == true,
+				format: i[2]&.dig(:format) || :plaintext,
 				disable_quotes: i[2]&.dig(:disable_quotes) == true,
 				disable_hyphens: i[2]&.dig(:disable_hyphens) == true,
 				disable_nbsp: i[2]&.dig(:disable_nbsp) == true,
-				disable_nobr: i[2]&.dig(:disable_nobr) == true
+				disable_nobr: i[2]&.dig(:disable_nobr) == true,
+				disable_escape: i[2]&.dig(:disable_escape) == true
 		)
-		f.gsub!(Lightgraf::Fixtures::NBSP, Lightgraf::Fixtures::SPACE) if ignore_nbsp
+		if ignore_nbsp
+			f.gsub!(Lightgraf::Fixtures::NBSP, Lightgraf::Fixtures::SPACE)
+			f.gsub!(Lightgraf::Fixtures::HTML_NBSP, Lightgraf::Fixtures::SPACE)
+		end
 		f.gsub!(replace[0], replace[1]) unless replace.nil?
 		assert_equal i[1], f
 	end